// src/shared/api/generated/model/searchProductsProductGroup.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export type SearchProductsProductGroup = typeof SearchProductsProductGroup[keyof typeof SearchProductsProductGroup];


export const SearchProductsProductGroup = {
  SAVINGS: 'SAVINGS',
  DEPOSIT: 'DEPOSIT',
} as const;