// src/shared/api/generated/model/searchProductsSort.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */

export type SearchProductsSort = typeof SearchProductsSort[keyof typeof SearchProductsSort];


export const SearchProductsSort = {
  RATE: 'RATE',
  NEW: 'NEW',
  NAME: 'NAME',
} as const;