// src/shared/api/generated/model/pageResponseProductListItemResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CoreBank API Specification
 * Corebank REST API 명세서
 * OpenAPI spec version: v1.0.0
 */
import type { ProductListItemResponse } from './productListItemResponse';

export interface PageResponseProductListItemResponse {
  page?: number;
  size?: number;
  totalCount?: number;
  totalPages?: number;
  items?: ProductListItemResponse[];
}